+++ website/frontend/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import HomePage from 'src/pages/HomePage';
-import CommunityPage from './src/pages/Solutions/CommunityPage';
+import CommunityPage from './src/pages/CommunityPage';
 import ResearchPage from 'src/pages/ResearchPage';
 import Press from 'src/pages/Press/Press';
